12. Print (safe version)

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -10,7 +10,7 @@ size_t print_listint_safe(const listint_t *head)
 {
 	long int dist;
 	listint_t *now;
-	int counter = 0;
+	size_t counter = 0;
 
 	now = (listint_t *)head;
